Type checking

+++ 043_eval1/pandemic.c
@@ -87,8 +87,8 @@ void calcRunningAvg(unsigned * data, size_t n_days, double * avg) {
   if (n_days < 7) {
     return;
   }
-  unsigned sevenDayTotal = 0;
-  unsigned * firstDay = data;
+  uint64_t sevenDayTotal = 0;
+  uint64_t * firstDay = data;
 
   size_t day;
   for (day = 0; day < 7; day++) {
